+++ test/unit-tests/input/scan_code_tests.cpp
@@ -90,7 +90,7 @@ TEST(ScanCode, StringAssignmentOperator)
 
 TEST(ScanCode, Count)
 {
-  ASSERT_EQ(static_cast<int>(SDL_NUM_SCANCODES), cen::scan_code::count());
+  ASSERT_EQ(SDL_NUM_SCANCODES, cen::scan_code::count());
 }
 
 TEST(ScanCode, Unknown)
